habitapp_backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
dj-database-url
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
